requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
langdetect==1.0.9
jinja2==3.1.4
python-multipart==0.0.9
numpy>=1.25.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
